Commented out the bag of feature implementation, and created a function called process which will run the bag of feature implementation. So far this function loads the images, extracts SIFT features, and clusters them to build the dictionary. So far, everything seems to work correctly.
Added the Dictionary class into the imagefeatures.h file, and created
.cpp files for the imagefeatures, histograms, objectset, and dictionary.
This should help keep everything organized and compact.

Things to do next: Add a BoF parameter class, which will hold all the
sets of parameters that are to be used for the BoF method (SIFT, clustering,
SVM, etc...). Uncomment out the SVM code, perhaps even create a class to
handle all that.